플레이어체력바UI+화면바깥총알삭제+스타트버튼수정+슬롯머신아이템/옷장아이템넣기

+++ StreetCatFighter/Assets/Scripts/BattleScript/Enemybullet.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class Enemybullet : MonoBehaviour//보스 공격총알 스프라이트
 {
     [SerializeField]
-    private int damage = 1;
+    private float damage = 1f;
 
     private void OnTriggerEnter2D(Collider2D collision)
     {
